+++ src/commands/base/CliCommandBase.ts
@@ -1,10 +1,14 @@
 import * as path from "path";
-import { spawn } from 'child_process';
+import * as os from "os";
+import { spawn, execSync } from 'child_process';
 import { TreeItem, ContextValues } from '../../tree';
 import { CommandBase } from './CommandBase';
 import { SolutionExplorerProvider } from '../../SolutionExplorerProvider';
+import * as iconv from 'iconv-lite';
 
 export abstract class CliCommandBase extends CommandBase {
+    private codepage: string = "65001";
+
     constructor(protected readonly provider: SolutionExplorerProvider, protected readonly app: string) {
         super();
     }
@@ -15,17 +19,18 @@ export abstract class CliCommandBase extends CommandBase {
     }
 
     protected runCliCommand(app: string, args: string[], path: string): Promise<void> {
+        this.checkCurrentEncoding();
         this.provider.logger.log('Cli: ' + [ app, ...args ].join(' '));
 
         return new Promise(resolve => {
             let process = spawn(app, args, { cwd: path });
             
-            process.stdout.on('data', (data: string) => {
-                this.provider.logger.log(data);
+            process.stdout.on('data', (data: Buffer) => {
+                this.provider.logger.log(this.decode(data));
             });
             
-            process.stderr.on('data', (data: string) => {
-                this.provider.logger.error(data);
+            process.stderr.on('data', (data: Buffer) => {
+                this.provider.logger.error(this.decode(data));
             });
             
             process.on('exit', (code) => {
@@ -40,5 +45,22 @@ export abstract class CliCommandBase extends CommandBase {
         if (item.project) return path.dirname(item.project.fullPath);
         if (item.solution) return item.solution.FolderPath;
         return null;
+    }
+
+    private checkCurrentEncoding(): void {
+        if (os.platform() === "win32") {
+            this.codepage = execSync('chcp').toString().split(':').pop().trim();
+        }
+    }
+
+    private decode(data: Buffer): string {
+        switch (this.codepage) {
+            case "932":
+                return iconv.decode(data, "Shift_JIS");
+            case "936":
+                return iconv.decode(data, "GBK");
+            default:
+                return data.toString();
+        }
     }
 }